remove extendable, mode -model, -models, proxies
