--- addon/properties/-model.js
@@ -1,10 +1,0 @@
-import createModel from './-create-model';
-
-export default createModel({
-  create(owner, opts, type, build) {
-    if(!type) {
-      return;
-    }
-    return build((target, parent, modelOpts) => target._createInternalModel(type, parent, modelOpts));
-  }
-});
