--- addon/properties/create-defaults-mixin.js
@@ -1,6 +1,0 @@
-export default opts => extendable => extendable.extend(opts_ => {
-  if(typeof opts === 'function') {
-    return opts(opts_);
-  }
-  return opts;
-});
